fix for various issues around image size and bundles

+++ app.js
@@ -131,7 +131,7 @@ app.post('/api/classify', app.upload.single('images_file'), function(req, res, n
     return async.reflect(visualRecognition[method].bind(visualRecognition, params));
   }), function(err, results) {
     // delete the recognized file
-    if (params.images_file) {
+    if (params.images_file && !req.body.url) {
       fs.unlink(params.images_file.path);
     }
